+++ libs/schema/src/index.ts
@@ -1,27 +1,28 @@
-export * from "./lib/schemaAttribute";
-export * from "./lib/schemaCommand";
-export * from "./lib/schemaConstant";
-export * from "./lib/schemaEnum";
-export * from "./lib/schemaEnumMember";
-export * from "./lib/schemaErrorCodes";
-export * from "./lib/schemaInterface";
-export * from "./lib/schemaOperation";
-export * from "./lib/schemaProperty";
-export * from "./lib/schemaQuery";
-export * from "./lib/schemaTopic";
+export * from "./lib/schemaAttribute"
+export * from "./lib/schemaCommand"
+export * from "./lib/schemaConstant"
+export * from "./lib/schemaEnum"
+export * from "./lib/schemaEnumMember"
+export * from "./lib/schemaErrorCodes"
+export * from "./lib/schemaExtensions"
+export * from "./lib/schemaInterface"
+export * from "./lib/schemaOperation"
+export * from "./lib/schemaProperty"
+export * from "./lib/schemaQuery"
+export * from "./lib/schemaTopic"
 
-export * from "./lib/types/schemaGenericType";
-export * from "./lib/types/schemaInternalType";
-export * from "./lib/types/schemaKnownType";
-export * from "./lib/types/schemaType";
+export * from "./lib/types/schemaGenericType"
+export * from "./lib/types/schemaInternalType"
+export * from "./lib/types/schemaKnownType"
+export * from "./lib/types/schemaType"
 
-export * from "./lib/values/schemaBooleanValue";
-export * from "./lib/values/schemaNullValue";
-export * from "./lib/values/schemaNumberValue";
-export * from "./lib/values/schemaStringValue";
-export * from "./lib/values/schemaValue";
+export * from "./lib/values/schemaBooleanValue"
+export * from "./lib/values/schemaNullValue"
+export * from "./lib/values/schemaNumberValue"
+export * from "./lib/values/schemaStringValue"
+export * from "./lib/values/schemaValue"
 
-export * from "./lib/parseSchema";
+export * from "./lib/parseSchema"
 
 // eslint-disable-next-line import/no-useless-path-segments
-export { leancode } from "./lib/protocol/index.js";
+export { leancode } from "./lib/protocol/index.js"