+++ packages/api-admin/src/lib/components/Table/filters/mkColumnRender.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import { ReactNode } from "react"
 import { AdminTableColumn, EnumsMap } from "@leancodepl/contractsgenerator-typescript-plugin-admin"
 import { defaultFormatters } from "../formatters/defaultFormatters"