+++ packages/api-admin/src/lib/components/Table/filters/rangeFilters/DateRangePickerFilter/index.tsx
@@ -3,10 +3,10 @@ import { FilterDropdownProps } from "antd/lib/table/interface"
 import dayjs from "dayjs"
 import { ConfirmResetButtons } from "../../common/ConfirmResetButtons"
 
-type DatePickerFilterProps = {
+type DatePickerFilterProps = FilterDropdownProps & {
     placeholder?: string
     type?: "dateOnly" | "dateTimeOffset"
-} & FilterDropdownProps
+}
 
 export function DateRangePickerFilter({
     setSelectedKeys,
@@ -18,7 +18,6 @@ export function DateRangePickerFilter({
     return (
         <div style={{ padding: 8, display: "flex", flexDirection: "column", gap: 8 }}>
             <DatePicker.RangePicker
-                // eslint-disable-next-line jsx-a11y/no-autofocus
                 autoFocus
                 showTime={type === "dateTimeOffset"}
                 value={[