+++ packages/api-admin/src/lib/components/Table/filters/singleValueFilters/DatePickerFilter/index.tsx
@@ -4,10 +4,10 @@ import { DatePicker } from "../../../../DatePicker"
 import { ConfirmResetButtons } from "../../common/ConfirmResetButtons"
 import { FilterDropdownWrapper } from "../../common/FilterDropdownWrapper"
 
-type DatePickerFilterProps = {
+type DatePickerFilterProps = FilterDropdownProps & {
     placeholder?: string
     type?: "dateOnly" | "dateTimeOffset"
-} & FilterDropdownProps
+}
 
 export function DatePickerFilter({
     setSelectedKeys,
@@ -19,11 +19,10 @@ export function DatePickerFilter({
     return (
         <FilterDropdownWrapper>
             <DatePicker
-                // eslint-disable-next-line jsx-a11y/no-autofocus
                 autoFocus
                 showTime={type === "dateTimeOffset"}
                 value={selectedKeys[0] ? dayjs(selectedKeys[0]) : null}
-                onChange={(_, date) => setSelectedKeys?.([date])}
+                onChange={(_, date) => setSelectedKeys?.(Array.isArray(date) ? date : [date])}
             />
             <ConfirmResetButtons clearFilters={clearFilters} confirm={confirm} />
         </FilterDropdownWrapper>