+++ packages/api-admin/src/lib/types/api.ts
@@ -1,5 +1,3 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-
 import { mkCqrsClient } from "@leancodepl/react-query-cqrs-client"
 
 type ContractParams = Record<string, any> | undefined
@@ -13,4 +11,4 @@ export type PoorMansAPI = {
 
 export type ClientFactoryContracts = Record<string, [ContractParams, ContractResponse]>
 
-export type CQRS = (cqrsClient: { createTopic: any } & ReturnType<typeof mkCqrsClient>) => Record<string, unknown>
+export type CQRS = (cqrsClient: ReturnType<typeof mkCqrsClient> & { createTopic: any }) => Record<string, unknown>