+++ packages/generator/src/generate.ts
@@ -1,31 +1,32 @@
+import { z } from "zod/v4"
 import { GeneratorSessionContext } from "@leancodepl/contractsgenerator-typescript-plugin"
 import { cache } from "./cache"
 import { generateFile } from "./generateFile"
 import { getSchemaCached } from "./getSchemaCached"
-import { contractsGeneratorConfigurationSchema } from "./schema"
 
-type ContractsGeneratorPluginOptions = Record<string, unknown>
+export type ContractsGeneratorPluginConfiguration = z.infer<typeof contractsGeneratorPluginConfigurationSchema>
 
-export type ContractsGeneratorPluginConfiguration = Record<string, ContractsGeneratorPluginOptions> | string
+const contractsGeneratorPluginOptionsSchema = z.record(z.string(), z.unknown())
 
-type ContractsGeneratorFileConfiguration = {
-    plugins: ContractsGeneratorPluginConfiguration[]
-    config?: ContractsGeneratorPluginOptions
-}
+const contractsGeneratorPluginConfigurationSchema = z.union([
+    z.record(z.string(), contractsGeneratorPluginOptionsSchema),
+    z.string(),
+])
 
-export type ContractsGeneratorConfiguration = {
-    config?: ContractsGeneratorPluginOptions
-    generates: Record<string, ContractsGeneratorFileConfiguration>
-}
+const contractsGeneratorFileConfigurationSchema = z.object({
+    plugins: z.array(contractsGeneratorPluginConfigurationSchema),
+    config: contractsGeneratorPluginOptionsSchema.optional(),
+})
+
+const contractsGeneratorConfigurationSchema = z.object({
+    config: contractsGeneratorPluginOptionsSchema.optional(),
+    generates: z.record(z.string(), contractsGeneratorFileConfigurationSchema),
+})
 
 export async function generate(unsafeConfig: unknown) {
     const config = contractsGeneratorConfigurationSchema.parse(unsafeConfig)
 
-    const sessionContext: GeneratorSessionContext = {
-        getSchema: getSchemaCached(cache),
-        metadata: {},
-        cache,
-    }
+    const sessionContext: GeneratorSessionContext = { getSchema: getSchemaCached(cache), metadata: {}, cache }
 
     const configl1 = config.config