+++ plugins/admin/src/index.ts
@@ -1,34 +1,37 @@
 import {
-  GeneratorPlugin,
-  GeneratorPluginInstance,
-  GeneratorSessionContext,
-} from "@leancodepl/contractsgenerator-typescript-plugin";
-import { adminGeneratorPluginConfigurationSchema } from "./configuration.validator";
-import { generateAdmin } from "./lib/generateAdmin";
-import { printConfig } from "./lib/printConfig";
+    GeneratorPlugin,
+    GeneratorPluginInstance,
+    GeneratorSessionContext,
+} from "@leancodepl/contractsgenerator-typescript-plugin"
+import { adminGeneratorPluginConfigurationSchema } from "./configuration"
+import { generateAdmin } from "./lib/generateAdmin"
+import { printConfig } from "./lib/printConfig"
 
 class AdminGeneratorPlugin implements GeneratorPluginInstance {
-  configuration;
+    configuration
 
-  constructor(unsafeConfig: unknown, private context: GeneratorSessionContext) {
-    this.configuration = adminGeneratorPluginConfigurationSchema.parse(unsafeConfig);
-  }
+    constructor(
+        unsafeConfig: unknown,
+        private context: GeneratorSessionContext,
+    ) {
+        this.configuration = adminGeneratorPluginConfigurationSchema.parse(unsafeConfig)
+    }
 
-  async generate(): Promise<string> {
-    const schema = await this.context.getSchema(this.configuration.input);
+    async generate(): Promise<string> {
+        const schema = await this.context.getSchema(this.configuration.input)
 
-    const admin = generateAdmin(schema, this.configuration.nameTransform ?? (id => id));
+        const admin = generateAdmin(schema, this.configuration.nameTransform ?? (id => id))
 
-    return printConfig(admin);
-  }
+        return printConfig(admin)
+    }
 }
 
 const adminGeneratorPlugin: GeneratorPlugin = {
-  instance(unsafeConfig, context) {
-    return new AdminGeneratorPlugin(unsafeConfig, context.session);
-  },
-};
+    instance(unsafeConfig, context) {
+        return new AdminGeneratorPlugin(unsafeConfig, context.session)
+    },
+}
 
-export default adminGeneratorPlugin;
+export default adminGeneratorPlugin
 
-export * from "./contract";
+export * from "./contract"